requirements.txt
mysql-connector>=2.1.0,<2.2.0
python-dateutil
redis
requests
